backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
motor==3.3.2
email-validator==2.1.0
cors==1.0.1
Pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
